Refactored mock_tests/ folder
