+++ tests/mock_tests/common/mock_hiredis.cpp
